requirements.txt
torch>=1.9.0
pennylane>=0.21.0
gym>=0.21.0
numpy<2.0
PyYAML>=6.0
mlflow>=1.20.0
rich